+++ features/omni-kit/protocols/morpho-blue/hooks/index.ts
@@ -1,2 +1,2 @@
-export * from './useMorphoOmniData'
+export * from './useMorphoData'
 export * from './useMorphoSidebarTitle'
