+++ features/omni-kit/protocols/morpho-blue/hooks/useMorphoData.ts
@@ -1,0 +1,36 @@
+import { useProductContext } from 'components/context/ProductContextProvider'
+import type { OmniProtocolHookProps } from 'features/omni-kit/types'
+import { useObservable } from 'helpers/observableHook'
+import { useMemo } from 'react'
+import { EMPTY } from 'rxjs'
+
+export function useMorphoData({ dpmPositionData, tokenPriceUSDData }: OmniProtocolHookProps) {
+  const { morphoPosition$ } = useProductContext()
+
+  const [morphoPositionData, morphoPositionError] = useObservable(
+    useMemo(
+      () =>
+        dpmPositionData && tokenPriceUSDData
+          ? morphoPosition$(
+              tokenPriceUSDData[dpmPositionData.collateralToken],
+              tokenPriceUSDData[dpmPositionData.quoteToken],
+              dpmPositionData,
+            )
+          : EMPTY,
+      [dpmPositionData, tokenPriceUSDData],
+    ),
+  )
+
+  const morphoPositionAggregatedData = {
+    auction: { test: '' },
+    history: [],
+  }
+
+  return {
+    data: {
+      aggregatedData: morphoPositionAggregatedData,
+      positionData: morphoPositionData,
+    },
+    errors: [morphoPositionError],
+  }
+}
