--- features/omni-kit/protocols/morpho-blue/state/index.ts
@@ -1,1 +1,0 @@
-export * from './MorphoOmniCustomStateProvider'
